move db/api.py to moderation/api.py
