+++ src/blueprint_admin.py
@@ -3,7 +3,7 @@ from quart import Blueprint, redirect, request, url_for
 from blueprint_auth import admin_required
 from configs import CONSTS, DbType
 from db import query_dict, Phg
-from db.api import (
+from moderation.api import (
     create_user,
     delete_user,
     edit_user,
