+++ src/blueprint_auth.py
@@ -5,7 +5,11 @@ from werkzeug.security import check_password_hash
 
 from captcha import MathCaptcha
 from configs import CONSTS
-from db.api import get_user_with_username, is_user_admin, is_user_moderator
+from moderation.api import (
+    get_user_with_username,
+    is_user_admin,
+    is_user_moderator,
+)
 from enums import AuthActions
 from forms import LoginForm
 from render import render_controller
