+++ src/forms.py
@@ -24,7 +24,7 @@ from wtforms.validators import (
 )
 
 from configs import CONSTS
-from db.api import get_user_with_username, is_correct_password
+from moderation.api import get_user_with_username, is_correct_password
 from enums import ReportCategory, ReportStatus, SearchMode, UserRole
 from posts.capcodes import Capcode
